[stripe] chore: refactor redact header logic
