+++ api/server/stripe/patch-fetch.js
@@ -2,6 +2,7 @@ const { logger } = require('@librechat/data-schemas');
 const nodeFetch = require('node-fetch');
 const stripeForwardedHeaders = require('./forwardedHeaders');
 const { ProxyAgent } = require('proxy-agent');
+const { redactHeader } = require('./utils');
 
 /**
  * Patches the global fetch to use node-fetch instead of undici
@@ -45,7 +46,7 @@ function fetchLike(url, options = {}) {
   options = stripeForwardedHeaders.attach(options);
 
   // Log the request
-  logger.info(`[Stripe:patchFetch] fetch request ${formatRequest(url, options)}`);
+  logger.info(`[Stripe:patchFetch] CANONICAL-FETCH-REQUEST ${formatRequest(url, options)}`);
 
   // Use node-fetch to perform the request
   return nodeFetch(url, options);
@@ -53,10 +54,6 @@ function fetchLike(url, options = {}) {
 
 module.exports = patchFetch;
 
-function redactValue(value) {
-  if (!value) return '';
-  return `[REDACTED ${value.length} CHARACTERS]`;
-}
 
 function formatRequest(url, options) {
   const fields = [];
@@ -64,7 +61,7 @@ function formatRequest(url, options) {
   fields.push(`url="${formatUrl(url)}"`);
   fields.push(`headers=${JSON.stringify(formatHeaders(options.headers || new Headers()))}`);
   if (options.body) {
-    fields.push(`body="${redactValue(options.body)}"`);
+    fields.push(`body="${redactHeader('body', options.body)}"`);
   }
   return fields.join(' ');
 }
@@ -72,7 +69,7 @@ function formatRequest(url, options) {
 function formatUrl(url) {
   const urlObj = new URL(url);
   urlObj.searchParams.forEach((value, key) => {
-    urlObj.searchParams.set(key, redactValue(value));
+    urlObj.searchParams.set(key, redactHeader(key, value));
   });
   return urlObj.toString();
 }
@@ -80,7 +77,7 @@ function formatUrl(url) {
 function formatHeaders(headers) {
   const obj = {};
   for (const [key, value] of headers) {
-    obj[key] = redactValue(value);
+    obj[key] = redactHeader(key, value);
   }
   return obj;
 }
